+++ src/entrypoints/sig.content/index.ts
@@ -6,7 +6,7 @@ import { createStudent, syncHistory } from '@/services/next';
 import { processingToast, errorToast, successToast } from '@/utils/toasts'
 import { calculateQuadrimestres } from '@/utils/season'
 import { sendMessage } from '@/messaging';
-import { getStudentSigHistory, type CompleteStudent } from '@/services/ufabc-parser';
+import { getStudentSigHistory } from '@/services/ufabc-parser';
 
 export default defineContentScript({
 	async main() {
@@ -29,6 +29,7 @@ export default defineContentScript({
 
         processingToast.showToast();
         const { data: currentStudent, error } = await scrapeMenu($trs, sessionId, viewStateID.value);
+        console.log(currentStudent)
         // TODO(Joabe): move to background processing
         // https://webext-core.aklinker1.io/proxy-service/installation/
         const history = await getStudentSigHistory(sessionId, viewStateID.value, 'history');
